requirements.txt
fastapi
uvicorn
pypdf
langchain
langchain-community
openai
transformers
python-dotenv
